requirements.txt
playwright==1.55.0
python-dotenv==1.1.1
bs4==0.0.2